scraper/requirements.txt
# Web scraping
playwright==1.40.0
beautifulsoup4==4.12.2
selectolax==0.3.17
httpx==0.25.2
aiohttp==3.9.1

# Database
sqlalchemy==2.0.23
mysql-connector-python==8.2.0
alembic==1.12.1

# Data processing
pydantic==2.5.0
pandas==2.1.4
numpy==1.25.2

# AI and notifications
openai==1.3.0
requests==2.31.0

# Environment
python-dotenv==1.0.0

# Async support
asyncio
